pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vowel"
version = "0.3.0-beta"
description = "A modular evaluation framework for testing functions with YAML-based specifications"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Mert", email = "contact@tomris.dev"},
]
keywords = ["testing", "evaluation", "yaml", "pydantic-evals"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "watchdog",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=1.0.0",
    "pydantic-evals>=1.0.0",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-pyyaml",
]
mcp = [
    "mcp",
]
logfire = [
    "logfire",
]
optimize = [
    "vowel-optimization"
]
all = [
    "vowel[optimize,dev,mcp,logfire]",
]

[project.scripts]
vowel = "vowel.cli:main"

[project.urls]
Homepage = "https://github.com/fswair/vowel"
Documentation = "https://github.com/fswair/vowel#readme"
Repository = "https://github.com/fswair/vowel.git"
Issues = "https://github.com/fswair/vowel/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/vowel"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.ty]
[tool.ty.environment]
python-version = "3.11"

[tool.ty.rules]
unresolved-import = "ignore"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--color=yes",
]
markers = [
    "yaml: tests that load YAML fixture files",
    "api: tests for RunEvals API variants",
    "evaluators: tests for specific evaluator types",
    "integration: integration tests",
    "llm: tests that require LLM API calls",
]
